Create Install-GOGGames.ps1
Official Upload
Something I've been working on for a CyberCafe. Thought I would share some of the Install Methods I use for GOG Installers. Generally, for the most part, it should work; however, depending on the age of the compiled installer from GOG, there are some odd errors, or the silent installers never work.
The paths and setup and more specifically for my use cases, but feel free to enjoy parts or all of the script.